+++ backend/requirements.txt
@@ -4,3 +4,4 @@ gunicorn
 requests
 grequests
 redis
+flask-cors
